Restructuring functions around datapack to split up into separate submodules (and move some more general functions to util)
