+++ openghg/util/__init__.py
@@ -115,5 +115,8 @@ from ._util import (
     unanimous,
     verify_site,
     verify_site_with_satellite,
+    check_unique,
+    find_repeats,
+    collate_strings,
 )
 from ._versions import show_versions, check_if_need_new_version
